% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alamak.R
\name{alamak}
\alias{alamak}
\title{alamak!}
\usage{
alamak(f, pixpal = "Jerry")
}
\arguments{
\item{f}{a function}

\item{pixpal}{a pixel pal, one of "Lenny" (supportive velociraptor), "buster"
(cool lemon), "Jerry" (a parrot that hates you) and "Oniji" (Edo period
Japanese kabuki actor who speaks in haiku only)}
}
\value{
the result of the function if there are no errors, otherwise
    it displays the pixel pal with the message. Warnings are also displayed but
    the result of the function is still returned.
}
\description{
Catches an error or warning and lets a pixel pal tell you all about it
}
\examples{

alamak(a + 4)

}
