% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixpals.R
\docType{data}
\name{compression}
\alias{compression}
\title{Code to create initial Pixel Pals
Sourced on loading}
\format{
An object of class \code{logical} of length 1.
}
\usage{
compression
}
\description{
Code to create initial Pixel Pals
Sourced on loading
}
\keyword{datasets}
