% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alamak.R
\name{compressPixPal}
\alias{compressPixPal}
\title{Compress a pixel pal}
\usage{
compressPixPal(pix)
}
\arguments{
\item{pix}{a character indicating a path to a .PNG file.}
}
\value{
returns a compressed Pixel Pal character matrix that occupies half the
    space in the terminal
}
\description{
Compresses a pixel pal using Unicode Blocks
}
\details{
Internal use only. Note: the Apple Terminal does not deal well with
    Unicode blocks. When the  package is loaded, if Apple Terminal is detected,
    compression is disabled.
}
\references{
kindly suggested in https://github.com/gdagstn/alamak/issues/2
    by Trevor L. Davis
}
