% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixpals.R
\docType{data}
\name{pixpals}
\alias{pixpals}
\title{Code to create initial Pixel Pals
Sourced on loading}
\format{
An object of class \code{list} of length 5.
}
\usage{
pixpals
}
\description{
Code to create initial Pixel Pals
Sourced on loading
}
\keyword{datasets}
