% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alamak.R
\name{alamak}
\alias{alamak}
\title{alamak!}
\usage{
alamak(f, pixpal = "Jerry", compress = TRUE)
}
\arguments{
\item{f}{a function}

\item{pixpal}{a Pixel Pal, one of "Lenny" (supportive velociraptor), "Buster"
(cool lemon), "Jerry" (a parrot that hates you), "Oniji" (Edo period
Japanese kabuki actor who speaks in haiku only), and "E10N" (a robot from
the future). Alternatively, a Pixel Pal can be created manually following
the instructions at \code{?makePixPal}.}

\item{compress}{logical, should the compressed Pixel Pals be used? Only valid
if using internal Pixel Pals specified above.}
}
\value{
the output of \code{f} if there are no errors, otherwise
    it displays the pixel pal with the (edited) message.
    Warnings are also displayed, but the output of the function is still
    returned.
}
\description{
Catches an error or warning and lets a pixel pal tell you all about it
}
\examples{

# Error message
alamak(a + 4)

# Warning message
alamak(as.numeric("alamak"))

# Change Pixel Pal
alamak(as.numeric("alamak"), "Lenny")

}
