% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alamak.R
\name{makePixPal}
\alias{makePixPal}
\title{Make a pixel pal}
\usage{
makePixPal(filepath)
}
\arguments{
\item{filepath}{a character indicating a path to a .PNG file.}
}
\value{
a character string converted to color ASCII through \code{crayon}
}
\description{
Generates a pixel pal from a PNG file
}
\examples{

fpath <- system.file("extdata", "buster.png", package="alamak")
buster = makePixPal(fpath)
cat(as.vector(t(buster)), sep = "")

## To make a new Pixel Pal to be used with `alamak()`

new_pixelpal = list("crayon" = makePixPal(fpath),
    "messages" = list(
      "Error" = c("The first possible message",
                  "The second possible message",
                  "And so on",
                  "You know the drill"),
      "Warning" = c("A warning message",
                   "Another warning message")
                    )
                   )

alamak(a + b, pixpal = new_pixelpal)

}
